srip0/requirements.txt
fastapi
uvicorn
sqlite3
pyjwt
